[ROCm] Fix FP8 platform-awareness in NSA module, device detection, and KV cache
- NSA module (quant_k_cache, dequant_k_cache, nsa_indexer, index_buf_accessor,
  triton_kernel, tilelang_kernel): Replace hardcoded torch.float8_e4m3fn with
  platform-aware fp8_dtype (e4m3fnuz on MI300X). Fix hardcoded 448.0 max value
  to use fp8_max (240.0 on MI300X).

- get_device_name(): Add gcnArchName-based fallback for ROCm setups where
  torch.cuda.get_device_name() returns empty string. Without this fix,
  MI300X-tuned Triton FP8 GEMM configs were never loaded, causing all
  block-scale FP8 GEMMs to use default (untuned) parameters.

- memory_pool.py: Add float8_e4m3fnuz to FP8 store_dtype and NSA FP8 checks.

- trtllm_fp8_kv_kernel.py: Make logical KV cache dtype platform-aware.

- quantization/utils.py: Use fp8_min/fp8_dtype for weight scale threshold and
  assertion checks.

+++ python/sglang/srt/layers/attention/nsa/triton_kernel.py
@@ -4,6 +4,8 @@ import torch
 import triton
 import triton.language as tl
 
+from sglang.srt.layers.quantization.fp8_kernel import fp8_dtype
+
 
 # Triton implementation
 @triton.jit
@@ -109,7 +111,7 @@ def act_quant(
     M = x_flat.size(0)
 
     # Allocate output tensors
-    y = torch.empty_like(x, dtype=torch.float8_e4m3fn)
+    y = torch.empty_like(x, dtype=fp8_dtype)
     y_flat = y.view(-1, N)
     s = x.new_empty(*x.size()[:-1], N // block_size, dtype=torch.float32)
     s_flat = s.view(-1, N // block_size)
